+++ src/main/java/com/example/adventureprogearjava/services/ProductReviewService.java
@@ -5,5 +5,5 @@ import com.example.adventureprogearjava.dto.ProductReviewDTO;
 import java.util.List;
 
 public interface ProductReviewService {
-    List<ProductReviewDTO> getAll(Long productId, Integer ratingFrom, Integer ratingTo);
+    List<ProductReviewDTO> getAll(Long productId, Double ratingFrom, Double ratingTo);
 }
